fix(dashboards): add project-level permission check for DiscoverSavedQueryDetailEndpoint (#72159)
If an organization user does not have project-level permission to a
project (for example, with Open Membership disabled), they should not be
allowed to do any actions with saved queries that are based on said
project.

+++ src/sentry/discover/endpoints/discover_saved_query_detail.py
@@ -55,6 +55,8 @@ class DiscoverSavedQueryDetailEndpoint(DiscoverSavedQueryBase):
         if not self.has_feature(organization, request):
             return self.respond(status=404)
 
+        self.check_object_permissions(request, query)
+
         return Response(serialize(query), status=200)
 
     def put(self, request: Request, organization, query) -> Response:
@@ -63,6 +65,8 @@ class DiscoverSavedQueryDetailEndpoint(DiscoverSavedQueryBase):
         """
         if not self.has_feature(organization, request):
             return self.respond(status=404)
+
+        self.check_object_permissions(request, query)
 
         try:
             params = self.get_filter_params(
@@ -97,6 +101,8 @@ class DiscoverSavedQueryDetailEndpoint(DiscoverSavedQueryBase):
         """
         if not self.has_feature(organization, request):
             return self.respond(status=404)
+
+        self.check_object_permissions(request, query)
 
         query.delete()
